### 2016-07-28 #45 @whedon
Submission: ChainConsumer
**Submitting author:** @samreay (<a href="http://orcid.org/0000-0003-2071-9349">Samuel Hinton</a>)
**Repository:** <a href="https://github.com/samreay/ChainConsumer" target ="_blank">https://github.com/samreay/ChainConsumer</a>
**Version:** 0.10.1
