### 2016-07-28 c2280a2
Adding smoothing to the contours as well, enabled by default. Changing smoothing to divide by a factor of two to stop overlarge kernels.
