### 2016-07-28 8758c96
Adding 2D Gaussian filtering
