+++ examples/customisations/plot_colours_shade.py
@@ -5,7 +5,7 @@ Colours and Shading
 
 Choose custom colours and plot multiple chains with shading.
 
-Normally when plotting more than one chain, shading is removed so
+Normally when plotting more than two chains, shading is removed so
 you can clearly see the outlines. However, you can turn shading back
 on and modify the shade opacity if you prefer colourful plots.
 
@@ -22,13 +22,15 @@ from chainconsumer import ChainConsumer
 if __name__ == "__main__":
     np.random.seed(2)
     cov = normal(size=(2, 2)) + np.identity(2)
-    data = multivariate_normal(normal(size=2), 0.5 * (cov + cov.T), size=100000)
+    d1 = multivariate_normal(normal(size=2), 0.5 * (cov + cov.T), size=100000)
     cov = normal(size=(2, 2)) + np.identity(2)
-    data2 = multivariate_normal(normal(size=2), 0.5 * (cov + cov.T), size=100000)
+    d2 = multivariate_normal(normal(size=2), 0.5 * (cov + cov.T), size=100000)
+    cov = normal(size=(2, 2)) + np.identity(2)
+    d3 = multivariate_normal(normal(size=2), 0.5 * (cov + cov.T), size=100000)
 
-    c = ChainConsumer().add_chain(data, parameters=["$x$", "$y$"]).add_chain(data2)
-    c.configure_general(colours=["#B32222", "#D1D10D"])
-    c.configure_contour(contourf=True, contourf_alpha=0.5)
+    c = ChainConsumer().add_chain(d1, parameters=["$x$", "$y$"]).add_chain(d2).add_chain(d3)
+    c.configure_general(colours=["#B32222", "#D1D10D", "#455A64"])
+    c.configure_contour(shade=True, shade_alpha=0.2)
     c.configure_bar(shade=True)
     fig = c.plot()
 
