+++ chainconsumer/chain.py
@@ -101,7 +101,7 @@ class ChainConsumer(object):
         return self
 
     def configure_general(self, bins=None, flip=True, rainbow=None, colours=None,
-                          serif=True, plot_hists=True, max_ticks=5, kde=False, smooth=10):  # pragma: no cover
+                          serif=True, plot_hists=True, max_ticks=5, kde=False, smooth=3):  # pragma: no cover
         r""" Configure the general plotting parameters common across the bar
         and contour plots. If you do not call this explicitly, the :func:`plot`
         method will invoke this method automatically.
@@ -158,7 +158,7 @@ class ChainConsumer(object):
         self.parameters_general["rainbow"] = rainbow
         self.parameters_general["plot_hists"] = plot_hists
         self.parameters_general["kde"] = kde
-        if not smooth:
+        if not smooth or kde:
             smooth = None
         self.parameters_general["smooth"] = smooth
         if colours is None:
@@ -169,8 +169,8 @@ class ChainConsumer(object):
         self._configured_general = True
         return self
 
-    def configure_contour(self, sigmas=None, cloud=None, contourf=None,
-                          contourf_alpha=1.0):  # pragma: no cover
+    def configure_contour(self, sigmas=None, cloud=None, shade=None,
+                          shade_alpha=None):  # pragma: no cover
         """ Configure the default variables for the contour plots. If you do not call this
         explicitly, the :func:`plot` method will invoke this method automatically.
 
@@ -185,10 +185,10 @@ class ChainConsumer(object):
             and [1, 2] for multiple chains.
         cloud : bool, optional
             If set, overrides the default behaviour and plots the cloud or not
-        contourf : bool, optional
+        shade : bool, optional
             If set, overrides the default behaviour and plots filled contours or not
-        contourf_alpha : float, optional
-            Filled contour alpha value override.
+        shade_alpha : float, optional
+            Filled contour alpha value override. Default is 1.0
         """
         num_chains = len(self.chains)
 
@@ -205,10 +205,15 @@ class ChainConsumer(object):
             cloud = False
         self.parameters_contour["cloud"] = cloud
 
-        if contourf is None:
-            contourf = num_chains == 1
-        self.parameters_contour["contourf"] = contourf
-        self.parameters_contour["contourf_alpha"] = contourf_alpha
+        if shade is None:
+            shade = num_chains <= 2
+        self.parameters_contour["shade"] = shade
+        if shade_alpha is None:
+            if num_chains == 1:
+                shade_alpha = 1.0
+            else:
+                shade_alpha = np.sqrt(1 / num_chains)
+        self.parameters_contour["shade_alpha"] = shade_alpha
 
         self._configured_contour = True
 
@@ -223,13 +228,15 @@ class ChainConsumer(object):
             Will not work if you have multiple chains
         shade : bool, optional
             If set to true, shades in confidence regions in under histogram. By default
-            this happens if you have a single chain, but is disabled if you are comparing
-            multiple chains.
+            this happens if you less than 3 chains, but is disabled if you are comparing
+            more chains.
         """
         if summary is not None:
             summary = summary and len(self.chains) == 1
         self.parameters_bar["summary"] = summary
-        self.parameters_bar["shade"] = shade if shade is not None else len(self.chains) == 1
+        if shade is None:
+            shade = len(self.chains) <= 2
+        self.parameters_bar["shade"] = shade
[…]
