+++ examples/customisations/plot_cloud_sigma.py
@@ -7,6 +7,10 @@ Choose custom sigma levels and display point cloud.
 
 In this example we display more sigma levels, turn on the point cloud, and
 disable the parameter summaries on the top of the marginalised distributions.
+
+Note that because of the very highly correlated distribution we have, it is
+useful to increase the number of bins the plots are generated with, to capture the
+thinness of the correlation.
 """
 
 import numpy as np
@@ -19,7 +23,7 @@ if __name__ == "__main__":
     data = multivariate_normal(normal(size=3), 0.5 * (cov + cov.T), size=100000)
 
     c = ChainConsumer().add_chain(data, parameters=["$x$", "$y$", "$z$"])
-    c.configure_bar(summary=False)
+    c.configure_bar(summary=False).configure_general(bins=1.4)
     c.configure_contour(cloud=True, sigmas=np.linspace(0, 2, 10))
     fig = c.plot()
 

[…]
