### 2016-07-28 514d49b
Updating travis because previous commands no longer work. Wonder why it has changed
