### 2016-07-28 10b3216
Disallowing sudo is really going to make things difficult
