### 2016-07-28 b7dd34e
Reverting to prior travis build and will raise an issue with travis in the morning if it is not fixed.
