### 2016-07-29 1e77cf9
Updating documentation to make it clearer
