### 2016-07-29 #45 @Samreay
Excellent, thanks Thomas. I had a request for a another feature, so the current version is 0.9.8 I believe. Hope that isn't an issue.
