### 2016-08-01 07fc4fa
Adding logging.basicConfig on initialisation
