### 2016-08-12 698ea01
Adding more style options, better extents and smarter tick labelling
