+++ examples/plot_introduction.py
@@ -8,6 +8,9 @@ A trivial example using data from a multivariate normal.
 We give truth values, parameter labels and set the figure size to
 fit one column of a two column document.
 
+It is important to note that (in future examples) we always call the
+configuration methods *after* loading in all the data.
+
 Note that the parameter summaries are all calculated from the chosen bin size and take
 into account if the data is being smoothed or not. It is thus important to consider
 whether you want smoothing enabled or (depending on your surfaces) more or less
