+++ chainconsumer/chain.py
@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import logging
 from scipy.interpolate import interp1d
-from matplotlib.ticker import MaxNLocator
+from matplotlib.ticker import MaxNLocator, ScalarFormatter
 import matplotlib.cm as cm
 import statsmodels.api as sm
 from scipy.ndimage.filters import gaussian_filter
@@ -14,7 +14,7 @@ class ChainConsumer(object):
     """ A class for consuming chains produced by an MCMC walk
 
     """
-    __version__ = "0.9.9"
+    __version__ = "0.9.10"
 
     def __init__(self):
         logging.basicConfig()
@@ -102,10 +102,15 @@ class ChainConsumer(object):
         return self
 
     def configure_general(self, bins=None, flip=True, rainbow=None, colours=None,
-                          serif=True, plot_hists=True, max_ticks=5, kde=False, smooth=3):  # pragma: no cover
+                          linestyles=None, linewidths=None, serif=True, plot_hists=True,
+                          max_ticks=5, kde=False, smooth=3):  # pragma: no cover
         r""" Configure the general plotting parameters common across the bar
         and contour plots. If you do not call this explicitly, the :func:`plot`
         method will invoke this method automatically.
+
+        Please ensure that you call this method *after* adding all the relevant data to the
+        chain consumer, as the consume changes configuration values depending on
+        the supplied data.
 
         Parameters
         ----------
@@ -121,9 +126,18 @@ class ChainConsumer(object):
             rotate the histogram so that it is horizontal.
         rainbow : bool, optional
             Set to True to force use of rainbow colours
-        colours : list[str(hex)], optional
+        colours : str(hex)|list[str(hex)], optional
             Provide a list of colours to use for each chain. If you provide more chains
-            than colours, you *will* get the rainbow colour spectrum.
+            than colours, you *will* get the rainbow colour spectrum. If you only pass
+            one colour, all chains are set to this colour. This probably won't look good.
+        linestyles : str, list[str], optional
+            Provide a list of line styles to plot the contours and marginalsied
+            distributions with. By default, this will become a list of solid lines. If a
+            string is passed instead of a list, this style is used for all chains.
+        linewidths : float, list[float], optional
+            Provide a list of line widths to plot the contours and marginalsied
+            distributions with. By default, this is a width of 1. If a float
+            is passed instead of a list, this width is used for all chains.
         serif : bool, optional
             Whether to display ticks and labels with serif font.
         plot_hists : bool, optional
@@ -163,10 +177,26 @@ class ChainConsumer(object):
             smooth = None
         self.parameters_general["smooth"] = smooth
         if colours is None:
-            self.parameters_general["colours"] = self.all_colours
+            if self.parameters_general.get("colours") is None:
+                self.parameters_general["colours"] = self.all_colours[:len(self.chains)]
         else:
+            if isinstance(colours, str):
+                colours = [colours] * len(self.chains)
             self.parameters_general["colours"] = colours
-
+        if linestyles is None:
+            if self.parameters_general.get("linestyles") is None:
+                self.parameters_general["linestyles"] = ["-"] * len(self.chains)
+        else:
+            if isinstance(linestyles, str):
+                linestyles = [linestyles] * len(self.chains)
+            self.parameters_general["linestyles"] = linestyles[:len(self.chains)]
+        if linewidths is None:
+            if self.parameters_general.get("linewidths") is None:
+                self.parameters_general["linewidths"] = [1.0] * len(self.chains)
+        else:
[…]
