### 2016-08-13 6f17832
Adding extra asserts
