### 2016-08-17 211eb41
Adding dictionary chain loading
