### 2016-08-25 #45 @tacaswell
- The current release on github is `0.10.0`, how can the submission be updated?
- I suggest making the installation instruction easier to find (give them a heading).  The setup.py should list `install_requires`.  Did not automatically install `statsmodels`
 - no clear link to contributing guidelines
 - statement of need in the documentation could be expanded (what does `corner` not do that this covers?).  The version in the paper is much better.
 - software archive missing from top of pdf
 - should deps other than mpl get citations?

 - what happens in `add_chain` with dict input + a list of parameters?

I recommend publication on the condition of addressing the above minor issues.
