### 2016-08-26 #45 @arfon
> The current release on github is 0.10.0, how can the submission be updated?

I can update this.

> I recommend publication on the condition of addressing the above minor issues.

Many thanks @tacaswell. @samreay - please let us know on this thread when you've had a chance to review and incorporate @tacaswell's feedback.
