+++ paper/paper.md
@@ -31,7 +31,8 @@ the functionality to plot the chains as a series of walks in
 parameter values, which provides an easy visual check on chain 
 mixing and chain convergence.
 
-Plotting is performed via the matplotlib library [@matplotlib]. 
+Plotting is performed via the matplotlib library [@matplotlib], and 
+makes use of various numpy [@numpy] and scipy [@scipy] functions.
 
 Code archives can be found on Zenodo at [@zenodo] and any
 bugs or feature requests can be opened as issues on the Github
