+++ paper/paper.bib
@@ -9,6 +9,32 @@
     doi = {10.1109/MCSE.2007.55}
 }
 
+@Misc{scipy,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
+  url = "http://www.scipy.org/"
+}
+
+@article{numpy,
+ author = {Walt, Stefan van der and Colbert, S. Chris and Varoquaux, Gael},
+ title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+ journal = {Computing in Science and Engg.},
+ issue_date = {March 2011},
+ volume = {13},
+ number = {2},
+ month = mar,
+ year = {2011},
+ issn = {1521-9615},
+ pages = {22--30},
+ numpages = {9},
+ url = {http://dx.doi.org/10.1109/MCSE.2011.37},
+ doi = {10.1109/MCSE.2011.37},
+ acmid = {1957466},
+ publisher = {IEEE Educational Activities Department},
+ address = {Piscataway, NJ, USA},
+ keywords = {NumPy, Python, Python, NumPy, scientific programming, numerical computations, programming libraries, numerical computations, programming libraries, scientific programming},
+}
 @online{github,
   author = {Samuel Hinton},
   title = {ChainConsumer},
@@ -22,5 +48,5 @@
   title = {ChainConsumer},
   doi = {10.5281/zenodo.58511},
   year = 2016,
-  howpublished = {\url{http://dx.doi.org/10.5281/zenodo.58511}}
+  howpublished = {\url{http://dx.doi.org/10.5281/zenodo.60315}}
 }
