+++ doc/chain_api.rst
@@ -5,6 +5,10 @@
 Chain Consumer API
 ==================
 
+.. autosummary:: chainconsumer.ChainConsumer
+    :members:
+    :undoc-members:
+
 .. autoclass:: chainconsumer.ChainConsumer
     :members:
     :undoc-members:
