### 2016-08-26 #45 @Samreay
Hi guys, am making the requested changes now. In regards to

> The setup.py should list install_requires. Did not automatically install statsmodels

Updated.

> no clear link to contributing guidelines

Have added a contribution heading with explicit details on the main page and the readme.

> statement of need in the documentation could be expanded

Have updated the introduction on the documentation page.

> software archive missing from top of pdf

Neither the example [in the about page](http://joss.theoj.org/about) nor the example [example page](https://github.com/arfon/fidgit/tree/master/paper) have the software archive shown at the top of the page - it is instead linked inline in the example. I also do not know how to generate the pdf itself, so am unsure what syntax to use. The identifier tag in the paper.json which contains the DOI is already filled out with the zenodo link, I thought that was going to be automatically injected.

> should deps other than mpl get citations?

Have added citations for numpy and scipy.

> what happens in add_chain with dict input + a list of parameters?

Have updated the API to be clear on this.
