+++ doc/chain_api.rst
@@ -7,6 +7,9 @@ Chain Consumer API
 
 .. toctree::
 
+    chainconsumer.ChainConsumer.add_chain
+
+
 .. autoclass:: chainconsumer.ChainConsumer
     :members:
     :undoc-members:
