### 2016-08-26 e3b042a
Giving up getting sphinx to do what I want
