+++ doc/chain_api.rst
@@ -5,10 +5,6 @@
 Chain Consumer API
 ==================
 
-.. toctree::
-
-    chainconsumer.ChainConsumer.add_chain
-
 
 .. autoclass:: chainconsumer.ChainConsumer
     :members:
