### 2016-08-27 #45 @arfon
> software archive missing from top of pdf

@tacaswell @Samreay - yes this is automatically injected at the end of the review once the submission review process is complete. Sorry for any confusion caused here.

Thanks for the rapid turnaround @Samreay. @tacaswell - would you mind taking another look to see if this is good to accept now?
