### 2016-08-28 #45 @tacaswell
Statsmodels is not directly cited, where as the other 3 direct dependencies are.  I defer to @arfon on if that is a blocking issue or not.

otherwise all of my concerns are addressend and looks good to publish :+1:

the version needs to be bumped to 0.10.1
