### 2016-08-28 #45 @Samreay
Yes, I hope to replace statsmodels with scikit-learns kde implementation in
their next release, so because of that, and statsmodels not having a
publicised bibtex citation (would have to @misc ref their github). In
addition, unlike matplotlib, numpy and scipy, which are used in all
instances of the library usage, the kde is only used on specific request
(which should be rare). For those reasons I left statsmodels out of the
citations. Would be happy to add them in, but would then have to update the
paper.md after the review when this changed, which seems also inappropriate.

Cheers

On Sun, Aug 28, 2016 at 1:05 PM, Thomas A Caswell <notifications@github.com>
wrote:

> Statsmodels is not directly cited, where as the other 3 direct
> dependencies are. I defer to @arfon <https://github.com/arfon> on if that
> is a blocking issue or not.
>
> otherwise all of my concerns are addressend and looks good to publish 👍
>
> the version needs to be bumped to 0.10.1
>
