### 2016-08-28 #45 @arfon
> Would be happy to add them in, but would then have to update the
paper.md after the review when this changed, which seems also inappropriate.

Provided this is the only change I think we should add Statsmodels to the current list of citations. @Samreay - could you please do this and then make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
