### 2016-08-29 a0ada00
Adding statsmodels
