+++ paper/paper.bib
@@ -12,7 +12,7 @@
 @Misc{scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
+  year =      {2001},
   url = "http://www.scipy.org/"
 }
 
@@ -49,4 +49,13 @@
   doi = {10.5281/zenodo.58511},
   year = 2016,
   howpublished = {\url{http://dx.doi.org/10.5281/zenodo.60315}}
+}
+
+@misc{statsmodels,
+  author = {{StatsModels Team}},
+  title = {StatsModels: Statistics in Python},
+  year = {2016},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/statsmodels/statsmodels}}
 }
