### 2016-08-29 #45 @Samreay
Have added statsmodels and archived. Final doi is
http://dx.doi.org/10.5281/zenodo.61064

On Mon, Aug 29, 2016 at 1:35 AM, Arfon Smith <notifications@github.com>
wrote:

> Would be happy to add them in, but would then have to update the
> paper.md after the review when this changed, which seems also
> inappropriate.
>
> Provided this is the only change I think we should add Statsmodels to the
> current list of citations. @Samreay <https://github.com/Samreay> - could
> you please do this and then make an archive of the reviewed software in
> Zenodo/figshare/other service and update this thread with the DOI of the
> archive? I can then move forward with accepting the submission.
>
