### 2016-08-30 #45 @arfon
OK thanks @Samreay. This is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00045 🚀 🎉 💥

Thanks for the review @tacaswell
