### 2016-08-30 #45 @Samreay
Thanks guys!

On Wed, Aug 31, 2016 at 1:18 AM, Arfon Smith <notifications@github.com>
wrote:

> Closed #45 <https://github.com/openjournals/joss-reviews/issues/45>.
>
